updated UI and clean code done with P0,P1&P2
