+++ slate-note-taking/src/Components/AuthForms/Login.css
@@ -1,17 +1,34 @@
 .Login-Form {
   display: flex;
-  justify-content: center;
   align-items: center;
-  background-color: silver;
-  border-radius: 0.3rem;
-  flex-direction: column;
-  padding: 3rem;
-  margin: 1rem;
-  height: auto;
-  width: 60%;
-  margin: 0 auto;
-  margin-top: 10rem;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
 }
-input {
-  margin: 1rem;
+
+.form-control {
+  display: flex;
+  height: auto;
+  align-items: center;
+  justify-content: center;
+  background-color: gray;
+  margin-top: 2rem;
+  box-shadow: 0px 0px 0px 1px rgba(0, 0, 0, 0.1);
+}
+.login-page-title {
+  font-size: 2rem;
+  font-weight: bold;
+  color: #000;
+  margin-top: 5rem;
+  text-align: center;
+}
+.login-instruction {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  padding: 1rem;
+}
+input:valid {
+  border: 1px solid green;
 }
