+++ slate-note-taking/src/Components/AuthForms/Signup.css
@@ -1,15 +1,33 @@
-@import "https://deploy-preview-13--agri-component.netlify.app/css/components.css";
 .Signup-Form {
   display: flex;
-  justify-content: center;
   align-items: center;
-  background-color: silver;
-  border-radius: 0.3rem;
-  flex-direction: column;
-  padding: 3rem;
-  margin: 1rem;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+}
+
+.form-control {
+  display: flex;
   height: auto;
-  width: 50%;
-  margin: 0 auto;
-  margin-top: 10rem;
+  align-items: center;
+  justify-content: center;
+  background-color: gray;
+  box-shadow: 0px 0px 0px 1px rgba(0, 0, 0, 0.1);
+}
+.login-page-title {
+  font-size: 2rem;
+  font-weight: bold;
+  color: #000;
+  margin-top: 5rem;
+  text-align: center;
+}
+.login-instruction {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 1rem;
+}
+input:valid {
+  border: 1px solid green;
 }
