+++ slate-note-taking/src/Components/EditForm/EditForm.css
@@ -56,9 +56,11 @@ input[type="text"].navigation__input {
 
 label.label-radio-box1 {
   height: 3rem;
-  width: 68.5%;
+  width: 46%;
   padding: 1rem;
-  margin-left: 5.7rem;
+  margin-left: 14.5rem;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
   display: flex;
   justify-content: center;
   align-items: center;
