+++ slate-note-taking/src/Components/Footer/Footer.css
@@ -11,6 +11,7 @@
   position: fixed;
   width: 100%;
   font-size: 1.2rem;
+  box-shadow: 0px -1px 0px 0px rgba(0, 0, 0, 0.1);
 }
 
 .social-media-links {
