+++ slate-note-taking/src/Components/Footer/Footer.js
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Footer.css";
 
 function Footer() {
