+++ slate-note-taking/src/Components/Hero/Hero.js
@@ -1,5 +1,4 @@
-import React from "react";
-import { Link } from "react-router-dom";
+import { Link } from "../../Utils/CustomUtils";
 import "./Hero.css";
 function Hero() {
   return (
