+++ slate-note-taking/src/Components/IndexAllComponents.js
@@ -1,0 +1,35 @@
+import Account from "./Accounts/Account";
+import Auth from "./Auth/Auth";
+import Login from "./AuthForms/Login";
+import Signup from "./AuthForms/Signup";
+import EditForm from "./EditForm/EditForm";
+import RTEEditor from "./Editor/RTEEditor";
+import Filters from "./FIlters/Filter";
+import Header from "./Header/Header";
+import Footer from "./Footer/Footer";
+import Hero from "./Hero/Hero";
+import InputNotes from "./InputNotes/InputNotes";
+import NotesModal from "./Modal/NotesModal";
+import NotesCard from "./NotesCard/Notescard";
+import TrashNotesCard from "./NotesCard/TrashNotesCard";
+import ArchiveNotesCard from "./NotesCard/ArchiveNotesCard";
+import Sidebar from "./Sidebar/Sidebar";
+
+export {
+  Account,
+  Auth,
+  Login,
+  Signup,
+  EditForm,
+  RTEEditor,
+  Filters,
+  Header,
+  Footer,
+  Hero,
+  InputNotes,
+  NotesModal,
+  NotesCard,
+  TrashNotesCard,
+  ArchiveNotesCard,
+  Sidebar,
+};
