+++ slate-note-taking/src/Components/InputNotes/InputNotes.css
@@ -55,6 +55,8 @@ label.label-radio-box {
   height: 3rem;
   width: 68.5%;
   padding: 1rem;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
   margin-left: 5.7rem;
   display: flex;
   justify-content: center;
@@ -89,6 +91,7 @@ label.label1-radio-box {
   align-items: center;
   justify-content: center;
   margin: 1rem;
+  width: 80%;
 }
 input[type="color"] {
   visibility: hidden;
