+++ slate-note-taking/src/Context/CoreContextFiles.js
@@ -1,0 +1,10 @@
+import ArchiveNotesContext from "./ArchiveNotesContext";
+import LoginSignupContext from "./LoginSignupContext";
+import NotetakingContext from "./NotetakingContext";
+import TrashNotesContext from "./TrashNotesContext";
+export {
+  ArchiveNotesContext,
+  LoginSignupContext,
+  NotetakingContext,
+  TrashNotesContext,
+};
