+++ slate-note-taking/src/Context/IndexAllContext.js
@@ -1,0 +1,11 @@
+import { useArchiveContext } from "./ArchiveNotesContext";
+import { useLoginSignupContext } from "./LoginSignupContext";
+import { useNoteTakingContext } from "./NotetakingContext";
+import { useTrashNotesContext } from "./TrashNotesContext";
+
+export {
+  useArchiveContext,
+  useLoginSignupContext,
+  useNoteTakingContext,
+  useTrashNotesContext,
+};
