+++ slate-note-taking/src/Pages/Accountpage.js
@@ -1,8 +1,10 @@
-import React from "react";
-import Account from "../Components/Accounts/Account";
-import Footer from "../Components/Footer/Footer";
-import Header from "../Components/Header/Header";
-import Sidebar from "../Components/Sidebar/Sidebar";
+import {
+  Account,
+  Footer,
+  Header,
+  Sidebar,
+} from "../Components/IndexAllComponents";
+
 function Accountpage() {
   return (
     <div>
