+++ slate-note-taking/src/Pages/IndexAllPages.js
@@ -1,0 +1,19 @@
+import Accountpage from "./Accountpage";
+import Archivespage from "./Archivespage";
+import Editnotespage from "./Editnotespage";
+import Homepage from "./Homepage";
+import Loginpage from "./Loginpage";
+import Signuppage from "./Signuppage";
+import Trashpage from "./Trashpage";
+import Welcomepage from "./Welcomepage";
+
+export {
+  Accountpage,
+  Archivespage,
+  Editnotespage,
+  Homepage,
+  Loginpage,
+  Signuppage,
+  Trashpage,
+  Welcomepage,
+};
