+++ slate-note-taking/src/Pages/Signuppage.js
@@ -1,7 +1,4 @@
-import React from "react";
-import Signup from "../Components/AuthForms/Signup";
-import Footer from "../Components/Footer/Footer";
-import Header from "../Components/Header/Header";
+import { Footer, Header, Signup } from "../Components/IndexAllComponents";
 
 function Signuppage() {
   return (
