+++ slate-note-taking/src/Pages/Welcomepage.js
@@ -1,14 +1,9 @@
-import React from "react";
-import Footer from "../Components/Footer/Footer";
-import Header from "../Components/Header/Header";
-import Hero from "../Components/Hero/Hero";
-import Sidebar from "../Components/Sidebar/Sidebar";
+import { Footer, Header, Hero } from "../Components/IndexAllComponents";
 
 function Welcomepage() {
   return (
     <div>
       <Header />
-      {/* <Sidebar /> */}
       <Hero />
       <Footer />
     </div>
